# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/": {
            "get": {
                "description": "Get basic service information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "general"
                ],
                "summary": "Root endpoint",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/email-import-status/{jobName}": {
            "get": {
                "description": "Gets the current status of an email import job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get email import job status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job name",
                        "name": "jobName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.JobStatus"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/import-emails": {
            "post": {
                "description": "Triggers a Kubernetes Job that downloads emails from Azure Blob Storage, imports them into PostgreSQL, and generates embeddings (all in one job)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Trigger end-to-end email import",
                "parameters": [
                    {
                        "description": "Import job parameters",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handlers.TriggerEmailImportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.TriggerEmailImportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.TriggerEmailImportResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.TriggerEmailImportResponse"
                        }
                    }
                }
            }
        },
        "/api/analytics": {
            "get": {
                "description": "Get analytics summary for a specified time period (today, yesterday, last_7_days, last_30_days)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get analytics summary",
                "parameters": [
                    {
                        "type": "string",
                        "default": "yesterday",
                        "description": "Time period (today, yesterday, last_7_days, last_30_days)",
                        "name": "period",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AnalyticsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.AnalyticsResponse"
                        }
                    }
                }
            }
        },
        "/api/analytics/daily-report": {
            "get": {
                "description": "Get analytics report for the previous day, suitable for daily Slack notifications",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get daily analytics report",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AnalyticsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.AnalyticsResponse"
                        }
                    }
                }
            }
        },
        "/api/chat": {
            "post": {
                "description": "Send a conversation to the AI chatbot and get a response with product recommendations enhanced by similar past conversations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Chat with AI using enhanced vector search (products + email history)",
                "parameters": [
                    {
                        "description": "Chat request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ChatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ChatResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ChatResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/models.ChatResponse"
                        }
                    }
                }
            }
        },
        "/api/chat/request-support": {
            "post": {
                "description": "Escalate conversation to support team with customer email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "support"
                ],
                "summary": "Request support escalation",
                "parameters": [
                    {
                        "description": "Support request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SupportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SupportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SupportResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SupportResponse"
                        }
                    }
                }
            }
        },
        "/api/healthz": {
            "get": {
                "description": "Get basic health status of the application",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.HealthResponse"
                        }
                    }
                }
            }
        },
        "/api/healthz/db": {
            "get": {
                "description": "Get database connectivity status and latency",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Database health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DBHealthResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/models.DBHealthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.JobStatus": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "integer"
                },
                "completion_time": {
                    "type": "string"
                },
                "failed": {
                    "type": "integer"
                },
                "job_name": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "succeeded": {
                    "type": "integer"
                }
            }
        },
        "handlers.TriggerEmailImportRequest": {
            "type": "object",
            "properties": {
                "source": {
                    "description": "Optional: specific source/path in blob storage",
                    "type": "string"
                }
            }
        },
        "handlers.TriggerEmailImportResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "job_name": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.AnalyticsResponse": {
            "description": "Analytics response payload",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": ""
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "summary": {
                    "$ref": "#/definitions/models.AnalyticsSummary"
                }
            }
        },
        "models.AnalyticsSummary": {
            "type": "object",
            "properties": {
                "email_embeddings_count": {
                    "description": "Emails processed for embeddings",
                    "type": "integer"
                },
                "email_embeddings_ran": {
                    "description": "Whether email embeddings ran in period",
                    "type": "boolean"
                },
                "email_threads": {
                    "description": "Total email threads",
                    "type": "integer"
                },
                "end_date": {
                    "description": "Period end",
                    "type": "string"
                },
                "openai_calls": {
                    "description": "Total OpenAI API calls",
                    "type": "integer"
                },
                "openai_tokens_used": {
                    "description": "Total tokens consumed",
                    "type": "integer"
                },
                "period": {
                    "description": "\"today\", \"yesterday\", \"last_7_days\", \"last_30_days\"",
                    "type": "string"
                },
                "product_embeddings_count": {
                    "description": "Products processed for embeddings",
                    "type": "integer"
                },
                "product_embeddings_ran": {
                    "description": "Embeddings info",
                    "type": "boolean"
                },
                "product_suggestions": {
                    "description": "Total product suggestions made",
                    "type": "integer"
                },
                "query_embeddings": {
                    "description": "Additional billing-relevant metrics",
                    "type": "integer"
                },
                "sendgrid_emails_sent": {
                    "description": "Emails sent via SendGrid",
                    "type": "integer"
                },
                "start_date": {
                    "description": "Period start",
                    "type": "string"
                },
                "support_escalations": {
                    "description": "Support requests sent",
                    "type": "integer"
                },
                "support_summarizations": {
                    "description": "GPT calls for support summaries (billable)",
                    "type": "integer"
                },
                "support_summary_tokens": {
                    "description": "Tokens used for support summarizations",
                    "type": "integer"
                },
                "thread_embeddings_count": {
                    "description": "Threads processed for embeddings",
                    "type": "integer"
                },
                "total_conversations": {
                    "description": "Total chat conversations",
                    "type": "integer"
                },
                "total_email_embeddings": {
                    "description": "Total email embeddings in DB",
                    "type": "integer"
                },
                "total_emails": {
                    "description": "Total emails in database",
                    "type": "integer"
                },
                "total_product_embeddings": {
                    "description": "Total product embeddings in DB",
                    "type": "integer"
                },
                "unique_products_viewed": {
                    "description": "Unique products suggested",
                    "type": "integer"
                }
            }
        },
        "models.ChatRequest": {
            "description": "Chat request payload",
            "type": "object",
            "properties": {
                "conversation": {
                    "description": "Array of conversation messages",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ConversationMessage"
                    }
                }
            }
        },
        "models.ChatResponse": {
            "description": "Chat response payload",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message if any",
                    "type": "string",
                    "example": ""
                },
                "products": {
                    "description": "Product name to SKU mapping for link generation",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "request_support": {
                    "description": "Whether to request customer email for support escalation",
                    "type": "boolean",
                    "example": false
                },
                "response": {
                    "description": "AI response message",
                    "type": "string",
                    "example": "Hello! How can I help you today?"
                }
            }
        },
        "models.ConversationMessage": {
            "description": "Single message in a conversation",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Message content",
                    "type": "string",
                    "example": "Hello, how can I help?"
                },
                "role": {
                    "description": "Message role (user, assistant, system)",
                    "type": "string",
                    "example": "user"
                }
            }
        },
        "models.DBHealthResponse": {
            "description": "Database health check response",
            "type": "object",
            "properties": {
                "connected": {
                    "description": "Database connection status",
                    "type": "boolean",
                    "example": true
                },
                "error": {
                    "description": "Error message if any",
                    "type": "string",
                    "example": ""
                },
                "latency": {
                    "description": "Database ping latency",
                    "type": "string",
                    "example": "1ms"
                },
                "status": {
                    "description": "Health status",
                    "type": "string",
                    "example": "healthy"
                },
                "timestamp": {
                    "description": "Timestamp of the check",
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
        "models.HealthResponse": {
            "description": "Health check response",
            "type": "object",
            "properties": {
                "status": {
                    "description": "Health status",
                    "type": "string",
                    "example": "healthy"
                },
                "timestamp": {
                    "description": "Timestamp of the check",
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "version": {
                    "description": "Application version",
                    "type": "string",
                    "example": "1.0.0"
                }
            }
        },
        "models.SupportRequest": {
            "description": "Support escalation request payload",
            "type": "object",
            "properties": {
                "conversation": {
                    "description": "Full conversation history",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ConversationMessage"
                    }
                },
                "customer_email": {
                    "description": "Customer email address",
                    "type": "string"
                }
            }
        },
        "models.SupportResponse": {
            "description": "Support escalation response payload",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message if any",
                    "type": "string",
                    "example": ""
                },
                "message": {
                    "description": "Response message",
                    "type": "string",
                    "example": "Email sent successfully"
                },
                "success": {
                    "description": "Whether the email was sent successfully",
                    "type": "boolean",
                    "example": true
                }
            }
        }
    }
}